major refactor for flask and imprt issues
