+++ const.py
@@ -1,0 +1,10 @@
+from enum import StrEnum
+
+class Const(StrEnum):
+    '''
+    Flask Consts
+    '''
+
+    UPLOAD_FOLDER = '/path/to/the/uploads'
+
+EXT = {'txt', 'pdf', 'csv'}
