+++ models/__init__.py
@@ -1,0 +1,3 @@
+'''
+init for dataclasses
+'''
