--- mydocky/agents/__init__.py
@@ -1,2 +1,0 @@
-from .agent import Agent
-from .registers import register_functions
