--- mydocky/chats/__init__.py
@@ -1,5 +1,0 @@
-'''
-init for chats
-'''
-
-from .similarty_search import Chat
