--- mydocky/models/__init__.py
@@ -1,4 +1,0 @@
-'''
-init for dataclasses
-'''
-from .dataclasses import OpenAIModels, Directories
