--- mydocky/utils/__init__.py
@@ -1,7 +1,0 @@
-'''
-init for utils package
-'''
-
-from .db_interface import Database
-from .document_loader import DocLoader
-from .embed_interface import EmbedDocs
