+++ templates/documentrender.html
@@ -1,0 +1,9 @@
+{% extends "template.html"%}
+
+{% block header %}
+<h1 class="display-5 fw-bolder text-white mb-2">Document Search Results:</h1>
+<div>
+        <md-block class="lead text-white-50 mb-4">{{ data }}.</md-block>
+</div>
+
+{% endblock %}
