+++ templates/documentsearch.html
@@ -1,0 +1,11 @@
+{% extends "template.html"%}
+
+{% block header %}
+<h1 class="display-5 fw-bolder text-white mb-2">Document Search</h1>
+<form action="/documentsearch" method="post">
+    <textarea class="commontextarea" name="docsearch" id="docsearch" rows="6" cols="70"></textarea>
+    <div>
+        <button type="submit">Submit</button>
+    </div>
+</form>
+{% endblock %}
