+++ templates/simplesearchrender.html
@@ -1,0 +1,9 @@
+{% extends "template.html"%}
+
+{% block header %}
+<h1 class="display-5 fw-bolder text-white mb-2">Search Results:</h1>
+<div>
+        <p class="lead text-white-50 mb-4">{{ data }}.</p>
+</div>
+
+{% endblock %}
