+++ templates/upload.html
@@ -1,0 +1,14 @@
+{% extends "template.html"%}
+
+{% block header %}
+<h1 class="display-5 fw-bolder text-white mb-2">Upload Document</h1>
+<form method=post enctype=multipart/form-data>
+    <div>
+        <label for="myfile">Select a file:</label>
+        <input id=myfile type=file name=file>
+    </div>
+    <div>
+        <input type=submit value=Upload>
+    </div>
+</form>
+{% endblock %}
